Do not keep global state for grpc_options
If grpc_options is not passed to __init__() a default list is
constructed. This argument is created once when the method is defined,
thus meaning all instances of gNMIclient ever created will share it. As
this list is later modified (e.g. in the case of skip_verify=True) all
instances share common options. This leads to connection failures and
can result in only one working GNMI connection per program launch.

For details on this type of error see also here:
https://docs.python-guide.org/writing/gotchas/#mutable-default-arguments