Wrap underlying exceptions instead of hiding them
In many places pygnmi catches an exception and then propagates a
Exception() with a generic error message to the library user. This is
probably done to tell the library user what the library tried to achieve
while something broke. In some cases, the library user needs access to
the original exception when trying to determine what specifically failed
(e.g. finding out if a call is retriable or not).

This commit replaces Exception with gNMIException in all places, where
an original exception would else be hidden. In some places we replace
the bare exception catching ("except:") with catching only Exception.
This was chosen instead of catching BaseException as we don't want to
catch exceptions like KeyboardInterrupt.