Replace Exception with less generic gNMIException
In many places the library raises a generic Exception with a small text.
To allow the user of the library to distinguish between explicitly
raised exceptions and potential programming errors or problems by
libraries used by pygnmi we now use gNMIException instead. As no parent
exception is caught, this parameter is left blank (i.e. no exception is
wrapped).