Unify exception behavior of get() and set()
get() and set() used to have a very different exception handling
behavior. get() would raise grpc exceptions, while all other exceptions
were suppressed and replaced by a generic error message without details.
set() would catch and return a gprc exception, but ignore (i.e. pass
through to the user) all normal / default exceptions.

get() and set() now both wrap and reraise grpc- and normal exceptions.
They are logged with an error message and reraised wrapped in a
gNMIException. This allows the user of the library to properly handle
either the gNMIException or the wrapped exception and also to e.g.
implement a custom retry logic based on what went wrong.

This is a breaking change in regards of error handling! Users who before
used the return value of set() now need to wrap this in a try-except
block (same as get()).

We also do not necessarily need to catch all Exceptions, but I went for
this in set() for the sake of consistency with get() and didn't want to
remove this from get(), as I don't know what else this method might
raise. We now only catch Exception instead of BaseException, as common
python best practice dictates.